Plumb the new int64 'EndOfFile' field through the rest of the backend and client

+++ lib/srv/desktop/rdp/rdpclient/src/lib.rs
@@ -759,7 +759,7 @@ pub struct CGOSharedDirectoryTruncateRequest {
     pub completion_id: u32,
     pub directory_id: u32,
     pub path: *const c_char,
-    pub end_of_file: u32,
+    pub end_of_file: i64,
 }
 
 pub type CGOSharedDirectoryTruncateResponse = SharedDirectoryTruncateResponse;
